# Conflicts:
#	backend/src/main/java/com/sm/backend/controller/OrderItemController.java
#	backend/src/main/java/com/sm/backend/controller/ProductController.java
#	backend/src/main/java/com/sm/backend/controller/ProductInventoryController.java
#	backend/src/main/java/com/sm/backend/controller/ProductVariantController.java

+++ backend/src/main/java/com/sm/backend/serviceImpl/ProductServiceImpl.java
@@ -69,13 +69,13 @@ public class ProductServiceImpl implements ProductService {
                    variant.setProduct(repository.findById(product.getProductId()).orElseThrow(() -> new ResourceNotFoundException("invalid product ID")));
                    variantRepository.save(variant);
 //            creating inventories
-                   ProductInventory inventory = new ProductInventory();
-                   inventory.setQuantity(request1.getInventoryRequest().getQuantity());
-                   inventory.setLocation(request1.getInventoryRequest().getLocation());
-                   inventory.setLastUpdated(request1.getInventoryRequest().getLastUpdated());
-                   inventory.setProduct(variant.getProduct());
-                   inventory.setProductVariant(variant);
-                   inventoryRepository.save(inventory);
+//                   ProductInventory inventory = new ProductInventory();
+//                   inventory.setQuantity(request1.getInventoryRequest().getQuantity());
+//                   inventory.setLocation(request1.getInventoryRequest().getLocation());
+//                   inventory.setLastUpdated(request1.getInventoryRequest().getLastUpdated());
+//                   inventory.setProduct(variant.getProduct());
+//                   inventory.setProductVariant(variant);
+//                   inventoryRepository.save(inventory);
                }
            }
         }
@@ -83,13 +83,13 @@ public class ProductServiceImpl implements ProductService {
     }
 
     @Override
-    public Object getById(Long productId) {
-        return repository.findById(productId)
-                .orElseThrow(() -> new ResourceNotFoundException("invalid Product ID"));
+    public ProductResponse getById(Long productId) {
+        return new ProductResponse(repository.findById(productId)
+                .orElseThrow(() -> new ResourceNotFoundException("invalid Product ID"))) ;
     }
 
     @Override
-    public Object getall(Integer pageNumber, Integer pageSize, String sortby, String sortDir) {
+    public List<ProductResponse> getAll(Integer pageNumber, Integer pageSize, String sortby, String sortDir) {
         Sort sort = null;
 
         if (sortDir.equalsIgnoreCase("asc")) {
@@ -131,18 +131,22 @@ public class ProductServiceImpl implements ProductService {
         Product product = repository.findById(productId).orElseThrow(() -> new ResourceNotFoundException("invalid Product ID"));
         List<ProductVariant> variants = variantRepository.getAllVariantsByProductId(product.getProductId());
 //        deleting inventories
-        for (ProductVariant variant:variants){
-        if (orderItemRepository.findOrderItemByProductVariant(variant).isPresent()) {
-        throw new ProductCanNotBeDeletedException("this product and its variants cannot be deleted since one of its variant is present in an order item");
-        }else{
-            inventoryRepository.delete(inventoryRepository.findProductInventoryByProductVariant(variant));
+       if (variants.isEmpty()){
+           repository.delete(product);
+       }
+           else {
+           for (ProductVariant variant : variants) {
+               if (orderItemRepository.findOrderItemByProductVariant(variant).isPresent()) {
+                   throw new ProductCanNotBeDeletedException("this product and its variants cannot be deleted since one of its variant is present in an order item");
+               } else {
+                   inventoryRepository.delete(inventoryRepository.findProductInventoryByProductVariant(variant));
 
 
-        }
+               }
 //        deleting variants
-        variantRepository.deleteAll(variants);
+               variantRepository.deleteAll(variants);
 //        deleting product
-        repository.delete(product);
-    }
-
+               repository.delete(product);
+           }
+       }
 }}
